ChatBot/requirements.txt
openai
azure-ai-generative[index,evaluate,promptflow]
azure-identity
langchain
langchain_community
langchain_openai
python-dotenv
pinecone-client
tiktoken